Simplify Module API
